Move RunReliably to netfx only
